+++ test/unit/specs/elements/GeoList/GeoListFooterButton.spec.js
@@ -1,11 +1,25 @@
 import { mount } from '@vue/test-utils'
 import GeoListFooterButton from '@/elements/GeoList/GeoListFooterButton.vue'
+import GeoTertiaryButton from '@/elements/GeoButton/GeoTertiaryButton.vue'
+import GeoButton from '@/elements/GeoButton/GeoButton.vue'
 
 describe('GeoListFooterButton', () => {
+  it('Should render GeoListFooterButton component', () => {
+    const wrapper = mount(GeoListFooterButton, {
+      stubs: {
+        'geo-tertiary-button': true
+      }
+    })
+    expect(wrapper.find('.geo-list-footer-button').exists()).toBe(true)
+  })
+
   it('Should render default slot', () => {
     const wrapper = mount(GeoListFooterButton, {
       slots: {
         default: [`<span class="more-results-footer-button-content">Load more results demo content</span>`]
+      },
+      stubs: {
+        'geo-tertiary-button': true
       }
     })
     expect(wrapper.find('.more-results-footer-button-content').exists()).toBe(true)
@@ -16,9 +30,22 @@ describe('GeoListFooterButton', () => {
     const wrapper = mount(GeoListFooterButton, {
       slots: {
         default: [`<span class="more-results-footer-button-content">Load more results demo content</span>`]
+      },
+      stubs: {
+        GeoTertiaryButton,
+        GeoButton
       }
     })
-    wrapper.find('.geo-list-footer-button__button').trigger('click')
+    wrapper.find(GeoTertiaryButton).trigger('click')
     expect(wrapper.emitted()['click']).toBeTruthy()
+  })
+
+  it('Should render a tertiary button inside it', () => {
+    const wrapper = mount(GeoListFooterButton, {
+      stubs: {
+        'geo-tertiary-button': true
+      }
+    })
+    expect(wrapper.find('geo-tertiary-button-stub').exists()).toBe(true)
   })
 })